requirements.txt
llmwhisperer-client==0.2.0
python-dotenv==1.0.1
